+++ src/SRISYS.Web/wwwroot/js/controllers/system/view-borrowers-controller.js
@@ -40,7 +40,9 @@
             vm.fetchBorrowers();
 		};
 
-		vm.delete = function (id) {
+        vm.delete = function (id) {
+            if (!utils.showConfirmMessage("Are you sure you want to delete this borrower?")) { return; }
+
 			borrowerService.deleteBorrower(id)
 				.then((response) => { vm.fetchBorrowers(); }, utils.onError)
 				.finally(utils.hideLoading);
